requirements.txt
sqlmodel
pandas
openpyxl
gspread
google-auth

